Implements #32 - additional storage providers Refactored the existing storage provider implementation and moved them all into dedicated nuget packages. Externalized the database model into a dedicated nuget package. Added additional PartTypes to the seed.
We will likely improve this much further later on, ideally we would instead write a file storage provider for EF Core and rely on storage providers already created for each platform. The current way of doing it requires maintaining queries across all packages and this is not good, but good for now.