Fixes #78 - No Binner.Web in release file v2.1.0
The issue was a problem with the build scripts after changing the target framework version.

While debugging this issue I discovered some other problems with the way the service was built and registered, so I refactored this out to prevent problems with duplicate IOC registrations and cleaned up unnecessary code.