Fixes #375 - label barcodes not scanning.
Updated the older barcode generation features for printing manual lables as the barcodes generated weren't scannable. Mostly this is due to Dymo ignoring the DPI settings and making it only available to their custom page settings UI, or through their very much broken SDK. Users will need to set it to 600dpi manually in Windows, I'm not even sure how this will be addressed in unix.